+++ Behaviours/MeltdownHandler.cs
@@ -1,6 +1,8 @@
-﻿using System.Collections;
+﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.Remoting.Metadata.W3cXsd2001;
+using FacilityMeltdown.API;
 using FacilityMeltdown.Behaviours;
 using FacilityMeltdown.Util;
 using GameNetcodeStuff;
@@ -12,24 +14,37 @@ using UnityEngine.Rendering.HighDefinition;
 
 namespace FacilityMeltdown {
     public class MeltdownHandler : MonoBehaviour {
-        PlayerControllerB player => GameNetworkManager.Instance.localPlayerController;
+        static PlayerControllerB Player => GameNetworkManager.Instance.localPlayerController;
 
-        private AudioSource musicSource, warningSound;
+        private AudioSource musicSource;
         internal static MeltdownHandler Instance;
 
         internal float meltdownTimer = 2 * 60;
         internal int causedBy;
 
         private bool radiationIncrease = false;
-        GameObject shockwave, explosion;
-        List<GameObject> spawnedParticleEffects = new List<GameObject>();
+        GameObject explosion;
+        List<MeltdownSequenceEffect> activeEffects = new List<MeltdownSequenceEffect>();
 
         Vector3 effectOrigin;
 
-        DialogueSegment[] shipTakeOffDialogue = new DialogueSegment[] { 
+        readonly static DialogueSegment[] shipTakeOffDialogue = new DialogueSegment[] { 
             new DialogueSegment { bodyText = "The company has deemed the current levels of radiation too high." },
             new DialogueSegment { bodyText = "The company can not risk damaging it's equipment." }
         };
+
+        readonly static DialogueSegment[] introDialogue = new DialogueSegment[] {
+                new DialogueSegment {
+                    bodyText = "... FAILED TO CONNECT TO INTERNAL FACILITY COMPUTER ... IDENTIFIYING ROOT CAUSE ..."
+                },
+                new DialogueSegment {
+                    bodyText = "UNSTABLE NUCLEAR REACTOR ... PREDICTING TIME UNTIL CATASTROPHIC EVENT ...",
+                    waitTime = 6
+                },
+                new DialogueSegment {
+                    bodyText = "2 MINUTES UNTIL CATASTRPOHIC NUCLEAR REACTOR EVENT"
+                }
+            };
 
         void Start() {
             if(Instance != null) {
@@ -45,15 +60,10 @@ namespace FacilityMeltdown {
             musicSource.loop = false;
             musicSource.Play();
 
-            warningSound = gameObject.AddComponent<AudioSource>();
-            warningSound.spatialBlend = 0;
-            warningSound.loop = false;
-
             // Effect Handlers
-            StartCoroutine(EmergencyLights());
-            StartCoroutine(EffectSpawningHandler());
-            StartCoroutine(WarningAnnouncerHandler());
-            StartCoroutine(ShockwaveSpawningHandler());
+            foreach(MeltdownSequenceEffect effect in MeltdownSequenceEffect.effects) {
+                StartCoroutine(PlayMeltdownSequenceEffect(effect));
+            }
 
             if (GameNetworkManager.Instance.localPlayerController.IsServer) {
                 List<EnemyVent> avaliableVents = new List<EnemyVent>();
@@ -71,112 +81,76 @@ namespace FacilityMeltdown {
                 }
             }
 
-            MeltdownPlugin.logger.LogInfo("Spawning Effects");
-            for(int i = 0; i < Random.Range(5, 15); i++) { // todo: make this scale based on map size
-                SpawnRandomEffect();
-            }
-
             effectOrigin = RoundManager.FindMainEntrancePosition(false, true);
             if(effectOrigin == Vector3.zero) {
                 MeltdownPlugin.logger.LogError("Effect Origin is Vector3.Zero! We couldn't find the effect origin");
                 HUDManager.Instance.DisplayGlobalNotification("Failed to find effect origin... Things will look broken.");
             }
 
-            HUDManager.Instance.ReadDialogue(new DialogueSegment[] {
-                new DialogueSegment {
-                    bodyText = "... FAILED TO CONNECT TO INTERNAL FACILITY COMPUTER ... IDENTIFIYING ROOT CAUSE ..."
-                },
-                new DialogueSegment {
-                    bodyText = "UNSTABLE NUCLEAR REACTOR ... PREDICTING TIME UNTIL CATASTROPHIC EVENT ...",
-                    waitTime = 6
-                },
-                new DialogueSegment {
-                    bodyText = "2 MINUTES UNTIL CATASTRPOHIC NUCLEAR REACTOR EVENT"
-                }
-            });
+            HUDManager.Instance.ReadDialogue(introDialogue);
             if (MeltdownConfig.Default.CFG_SCREEN_SHAKE.Value) {
                 HUDManager.Instance.ShakeCamera(ScreenShakeType.VeryStrong);
                 HUDManager.Instance.ShakeCamera(ScreenShakeType.Long);
             }
         }
-
-        Vector3 GetRandomPositionNearPlayer(float radius = 15f) {
-            Vector3 position = player.transform.position + (Random.insideUnitSphere * radius);
-            return RoundManager.Instance.GetRandomNavMeshPositionInBoxPredictable(position, 10f, layerMask: -1, randomSeed: new System.Random());
-        }
-
-        Vector3 GetRandomPositionInsideFacility() {
-            Vector3 position = RoundManager.Instance.insideAINodes[Random.Range(0, RoundManager.Instance.insideAINodes.Length)].transform.position;
-            return RoundManager.Instance.GetRandomNavMeshPositionInBoxPredictable(position, 10f, layerMask: -1, randomSeed: new System.Random());
-        }
-
-        void SpawnRandomEffect() {
-            Vector3 position = GetRandomPositionInsideFacility() + Vector3.up;
-            RaycastHit hit;
-            if (Physics.Raycast(new Ray(position, Vector3.up), out hit, 20f, 256)) {
-                GameObject prefab = Assets.facilityEffects[Random.Range(0, Assets.facilityEffects.Length)];
-                GameObject created = Instantiate(prefab);
-                prefab.transform.position = hit.point;
-                spawnedParticleEffects.Add(created);
-            }
-            MeltdownPlugin.logger.LogWarning("Failed to spawn effect, raycast failed.");
-        }
-
-
-        void CreateShockwave() {
-            if(shockwave != null) Destroy(shockwave);
-
-            shockwave = Instantiate(Assets.shockwavePrefab);
-            shockwave.transform.position = effectOrigin;
-            shockwave.AddComponent<Shockwave>();
-        }
         
         void OnDisable() { 
+            Instance = null;
             if(explosion != null)
                 Destroy(explosion);
-            if(shockwave != null)
-                Destroy(shockwave);
-            Instance = null;
 
-            foreach(GameObject effect in spawnedParticleEffects) {
-                Destroy(effect);
+            foreach(MeltdownSequenceEffect effect in activeEffects) {
+                try {
+                    effect.Cleanup();
+                } catch(Exception e) {
+                    MeltdownPlugin.logger.LogError(effect.FullName + " produced a " + e.GetType().Name + " during Cleanup(). Some objects may still be visible between moons." + "\n" + e);
+                }
             }
         }
 
-        IEnumerator EmergencyLights() {
+        IEnumerator PlayMeltdownSequenceEffect(MeltdownSequenceEffect effect) {
             yield return null;
-            yield return new WaitForSeconds(5f);
-
-            for (int i = 0; i < RoundManager.Instance.allPoweredLights.Count; i++) {
-                RoundManager.Instance.allPoweredLights[i].color = Color.red;
+            if (!effect.IsEnabledOnThisMoon(StartOfRound.Instance.currentLevel)) {
+                MeltdownPlugin.logger.LogInfo(effect.Name + " will not be playing on this moon.");
+                yield break;
             }
-
-            while (!HasExplosionOccured()) {
-                for (int i = 0; i < RoundManager.Instance.allPoweredLightsAnimators.Count; i++) {
-                    RoundManager.Instance.allPoweredLightsAnimators[i].SetBool("on", true);
-                }
-
-                yield return new WaitForSeconds(2f);
-
-                for (int i = 0; i < RoundManager.Instance.allPoweredLightsAnimators.Count; i++) {
-                    RoundManager.Instance.allPoweredLightsAnimators[i].SetBool("on", false);
-                }
-
-                yield return new WaitForSeconds(5f);
+            try {
+                effect.Setup();
+            } catch (Exception e) {
+                MeltdownPlugin.logger.LogError(effect.FullName + " produced a " + e.GetType().Name + " during Setup(). This error is fatal and the effect will not continue." + "\n" + e);
+                yield break;
             }
+            if (!effect.IsOneShot && !effect.Playing) {
+                MeltdownPlugin.logger.LogWarning(
+                    effect.FullName + " is not playing and is not marked as one shot. By default it will play once. If you are not the dev please report this." + 
+                    "If you are the dev, did you intend this effect to play once? Or play continously?" + 
+                    "If you meant to have it play once. Please overwrite .IsOneShot to be true." + 
+                    "If you meant to have it play continously. Please call base.Setup() in your Setup function."
+                    );
+            }
+            bool effectivelyOneShot = effect.IsOneShot || !effect.Playing;
+
+            if (effectivelyOneShot) {
+                yield return StartCoroutine(effect.Play(meltdownTimer));
+            } else {
+                while (!HasExplosionOccured() && effect.Playing) {
+                    yield return StartCoroutine(effect.Play(meltdownTimer));
+                }
+            }
+            
+            yield return StartCoroutine(effect.Stop());
 
             yield break;
         }
 
         void Update() {
+            if (HasExplosionOccured()) return;
             StartOfRound shipManager = StartOfRound.Instance;
 
             musicSource.volume = (float)MeltdownConfig.Default.CFG_MUSIC_VOLUME.Value / 100f;
-            warningSound.volume = (float)MeltdownConfig.Default.CFG_MUSIC_VOLUME.Value / 100f;
 
-            if (!player.isInsideFactory && !MeltdownConfig.Default.CFG_MUSIC_PLAYS_OUTSIDE.Value) {
+            if (!Player.isInsideFactory && !MeltdownConfig.Default.CFG_MUSIC_PLAYS_OUTSIDE.Value) {
                 musicSource.volume = 0;
-                warningSound.volume = 0;
             }
 
             meltdownTimer -= Time.deltaTime;
@@ -192,66 +166,11 @@ namespace FacilityMeltdown {
 
             if (meltdownTimer <= 0) {
                 musicSource.Stop();
-                warningSound.Stop();
 
                 explosion = GameObject.Instantiate(Assets.facilityExplosionPrefab);
                 explosion.transform.position = effectOrigin;
+                explosion.AddComponent<FacilityExplosionHandler>();
             }
-        }
-
-        IEnumerator WarningAnnouncerHandler() {
-            yield return null;
-
-            while (!HasExplosionOccured()) {
-                AudioClip sound = Assets.warnings[Random.Range(0, Assets.warnings.Length)];
-                warningSound.clip = sound;
-                warningSound.Play();
-
-                yield return new WaitForSeconds(Random.Range(10f, 17f));
-            }
-
-            yield break;
-        }
-
-        IEnumerator ShockwaveSpawningHandler() {
-            yield return null;
-
-            while (!HasExplosionOccured()) {
-                if (shockwave != null) Destroy(shockwave);
-
-                shockwave = Instantiate(Assets.shockwavePrefab);
-                shockwave.transform.position = effectOrigin;
-                shockwave.AddComponent<Shockwave>();
-
-
-                yield return new WaitForSeconds(Random.Range(20f, 30f));
-            }
-
-            yield break;
-        }
-
-        IEnumerator EffectSpawningHandler() {
-            yield return null;
-
-            while(!HasExplosionOccured()) {
-                for (int i = 0; i < Random.Range(5, 15); i++) { // todo: make this scale based on map size
-                    SpawnRandomEffect();
-                }
-
-                if (player.isInsideFactory)
-                    Instantiate(StartOfRound.Instance.explosionPrefab, GetRandomPositionNearPlayer(), Quaternion.Euler(-90f, 0f, 0f), RoundManager.Instance.mapPropsContainer.transform);
-                if (MeltdownConfig.Default.CFG_SCREEN_SHAKE.Value) {
-                    if (meltdownTimer > 60) {
-                        HUDManager.Instance.ShakeCamera(ScreenShakeType.Big);
-                    } else {
-                        HUDManager.Instance.ShakeCamera(ScreenShakeType.VeryStrong);
-                    }
-                }
-
-                yield return new WaitForSeconds(Random.Range(8f, 14f));
-            }
-
-            yield break;
         }
 
         IEnumerator ShipTakeOff() {
@@ -264,7 +183,7 @@ namespace FacilityMeltdown {
             yield return new WaitForSeconds(3f); // Wait for explosion
             yield return new WaitForSeconds(3f);
             HUDManager.Instance.shipLeavingEarlyIcon.enabled = false;
-            StartMatchLever startMatchLever = Object.FindObjectOfType<StartMatchLever>();
+            StartMatchLever startMatchLever = GameObject.FindObjectOfType<StartMatchLever>();
             startMatchLever.triggerScript.animationString = "SA_PushLeverBack";
             startMatchLever.leverHasBeenPulled = false;
             startMatchLever.triggerScript.interactable = false;